+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Releases
 
 * Unreleased
+    - New features:
+        - Dashboard now has update CSV export. #2249
     - Front end improvements:
         - Clearer relocation options while you’re reporting a problem #2238
     - Bugfixes: